Moved models to the /models folder according to Laravel standards (should be decided when defining the project standards).
