+++ database/factories/ClientFactory.php
@@ -2,7 +2,7 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Client;
+use App\Models\Client;
 use Faker\Generator as Faker;
 
 $factory->define(Client::class, function (Faker $faker) {
